requirements.txt
# trend-radar requirements
# Python 3.10+

# Data
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
lxml>=5.0.0
html5lib>=1.1

# Google Trends (Phase 3)
pytrends>=4.9.0

# Claude API (Phase 3)
anthropic>=0.39.0

# LINE API
requests>=2.31.0

# Dashboard (Phase 5)
streamlit>=1.30.0
plotly>=5.18.0

# Environment
python-dotenv>=1.0.0
